+++ app/__init__.py
@@ -122,10 +122,12 @@ def create_app() -> Flask:
     # Attach routes
     from app.common.auth import auth_blueprint
     from app.deliver_grant_funding.routes import deliver_grant_funding_blueprint
+    from app.developers.routes import developers_blueprint
     from app.healthcheck import healthcheck_blueprint
 
     app.register_blueprint(healthcheck_blueprint)
     app.register_blueprint(deliver_grant_funding_blueprint)
+    app.register_blueprint(developers_blueprint)
     app.register_blueprint(auth_blueprint)
 
     @app.route("/", methods=["GET"])
