+++ app/deliver_grant_funding/templates/deliver_grant_funding/developers/add_form.html
@@ -13,7 +13,7 @@
   {{
     govukBackLink({
         "text": "Back",
-        "href": url_for("deliver_grant_funding.add_form", grant_id = grant.id, schema_id=schema.id, section_id=section.id)
+        "href": url_for("developers.add_form", grant_id = grant.id, schema_id=schema.id, section_id=section.id)
     })
   }}
 {% endblock beforeContent %}
