+++ app/deliver_grant_funding/templates/deliver_grant_funding/developers/add_question.html
@@ -12,7 +12,7 @@
   {{
     govukBackLink({
         "text": "Back",
-        "href": url_for("deliver_grant_funding.choose_question_type", grant_id = grant.id, schema_id=schema.id, section_id=section.id, form_id=db_form.id, question_data_type=chosen_question_data_type.name, back_link="manage_section")
+        "href": url_for("developers.choose_question_type", grant_id = grant.id, schema_id=schema.id, section_id=section.id, form_id=db_form.id, question_data_type=chosen_question_data_type.name, back_link="manage_section")
     })
   }}
 {% endblock beforeContent %}
