+++ app/deliver_grant_funding/templates/deliver_grant_funding/developers/add_schema.html
@@ -12,7 +12,7 @@
   {{
     govukBackLink({
         "text": "Back",
-        "href": url_for("deliver_grant_funding.grant_developers_schemas", grant_id = grant.id)
+        "href": url_for("developers.grant_developers_schemas", grant_id = grant.id)
     })
   }}
 {% endblock beforeContent %}
