+++ app/deliver_grant_funding/templates/deliver_grant_funding/developers/edit_question.html
@@ -14,7 +14,7 @@
   {{
     govukBackLink({
         "text": "Back",
-        "href": url_for("deliver_grant_funding.manage_form", grant_id = grant.id, schema_id=schema.id, section_id=section.id, form_id=db_form.id, back_link="manage_section")
+        "href": url_for("developers.manage_form", grant_id = grant.id, schema_id=schema.id, section_id=section.id, form_id=db_form.id, back_link="manage_section")
     })
   }}
 {% endblock beforeContent %}
