+++ app/deliver_grant_funding/templates/deliver_grant_funding/developers/edit_schema.html
@@ -12,7 +12,7 @@
   {{
     govukBackLink({
       "text": "Back",
-      "href": url_for("deliver_grant_funding.manage_schema", grant_id=grant.id, schema_id=schema.id)
+      "href": url_for("developers.manage_schema", grant_id=grant.id, schema_id=schema.id)
     })
   }}
 {% endblock %}
