+++ app/deliver_grant_funding/templates/deliver_grant_funding/developers/edit_section.html
@@ -12,7 +12,7 @@
   {{
     govukBackLink({
       "text": "Back",
-      "href": url_for("deliver_grant_funding.manage_section", grant_id=grant.id, schema_id=schema.id, section_id = section.id)
+      "href": url_for("developers.manage_section", grant_id=grant.id, schema_id=schema.id, section_id = section.id)
     })
   }}
 {% endblock %}
