+++ app/deliver_grant_funding/templates/deliver_grant_funding/developers/select_form_type.html
@@ -14,7 +14,7 @@
   {{
     govukBackLink({
         "text": "Back",
-        "href": url_for("deliver_grant_funding.manage_section", grant_id = grant.id, schema_id=schema.id, section_id=section.id)
+        "href": url_for("developers.manage_section", grant_id = grant.id, schema_id=schema.id, section_id=section.id)
     })
   }}
 {% endblock beforeContent %}
@@ -51,7 +51,7 @@
         govukButton({
         "text": "Continue",
         "classes": "govuk-button--secondary",
-        "href": url_for("deliver_grant_funding.add_form", grant_id = grant.id, schema_id=schema.id, section_id=section.id, form_type="empty")
+        "href": url_for("developers.add_form", grant_id = grant.id, schema_id=schema.id, section_id=section.id, form_type="empty")
         })
       }}
     </div>
