+++ app/deliver_grant_funding/templates/deliver_grant_funding/manage_grant_base.html
@@ -19,7 +19,7 @@
       {
           "text": "Developers",
           "key": "grant_developers",
-          "href": url_for("deliver_grant_funding.grant_developers", grant_id=grant.id)
+          "href": url_for("developers.grant_developers", grant_id=grant.id)
       }
   ]
 %}
