+++ tests/integration/deliver_grant_funding/test_routes.py
@@ -115,7 +115,7 @@ def test_create_grant(authenticated_client, db_session):
 def test_create_collection_get(authenticated_client, factories, templates_rendered):
     grant = factories.grant.create()
     result = authenticated_client.get(
-        url_for("deliver_grant_funding.setup_schema", grant_id=grant.id),
+        url_for("developers.setup_schema", grant_id=grant.id),
     )
     assert result.status_code == 200
 
@@ -127,11 +127,11 @@ def test_create_collection_post(authenticated_client, factories, db_session):
     grant = factories.grant.create()
     collection_form = SchemaForm(name="My test collection")
     result = authenticated_client.post(
-        url_for("deliver_grant_funding.setup_schema", grant_id=grant.id),
+        url_for("developers.setup_schema", grant_id=grant.id),
         data=collection_form.data,
     )
     assert result.status_code == 302
-    assert result.location == url_for("deliver_grant_funding.grant_developers_schemas", grant_id=grant.id)
+    assert result.location == url_for("developers.grant_developers_schemas", grant_id=grant.id)
 
     grant_from_db = db_session.scalars(select(Grant).where(Grant.id == grant.id)).one()
     assert len(grant_from_db.collection_schemas) == 1
@@ -142,7 +142,7 @@ def test_create_collection_post_duplicate_name(authenticated_client, factories, 
     factories.collection_schema.create(name="My test collection", grant=grant)
     collection_form = SchemaForm(name="My test collection")
     result = authenticated_client.post(
-        url_for("deliver_grant_funding.setup_schema", grant_id=grant.id),
+        url_for("developers.setup_schema", grant_id=grant.id),
         data=collection_form.data,
     )
     assert result.status_code == 200
@@ -156,7 +156,7 @@ def test_create_collection_post_duplicate_name(authenticated_client, factories, 
 def test_create_section_get(authenticated_client, factories, templates_rendered):
     schema = factories.collection_schema.create()
     result = authenticated_client.get(
-        url_for("deliver_grant_funding.add_section", grant_id=schema.grant.id, schema_id=schema.id),
+        url_for("developers.add_section", grant_id=schema.grant.id, schema_id=schema.id),
     )
     assert result.status_code == 200
 
@@ -168,13 +168,11 @@ def test_create_section_post(authenticated_client, factories, db_session):
     schema = factories.collection_schema.create()
     section_form = SectionForm(title="My test section")
     result = authenticated_client.post(
-        url_for("deliver_grant_funding.add_section", grant_id=schema.grant.id, schema_id=schema.id),
+        url_for("developers.add_section", grant_id=schema.grant.id, schema_id=schema.id),
         data=section_form.data,
     )
     assert result.status_code == 302
-    assert result.location == url_for(
-        "deliver_grant_funding.list_sections", grant_id=schema.grant.id, schema_id=schema.id
-    )
+    assert result.location == url_for("developers.list_sections", grant_id=schema.grant.id, schema_id=schema.id)
 
     collection_from_db = db_session.scalars(select(CollectionSchema).where(CollectionSchema.id == schema.id)).one()
     assert len(collection_from_db.sections) == 1
@@ -185,7 +183,7 @@ def test_create_section_post_duplicate_name(authenticated_client, factories, db_
     factories.section.create(title="My test section", collection_schema=schema)
     section_form = SectionForm(title="My test section")
     result = authenticated_client.post(
-        url_for("deliver_grant_funding.add_section", grant_id=schema.grant.id, schema_id=schema.id),
+        url_for("developers.add_section", grant_id=schema.grant.id, schema_id=schema.id),
         data=section_form.data,
     )
 
@@ -199,7 +197,7 @@ def test_create_form_get(authenticated_client, factories, templates_rendered):
     section = factories.section.create()
     result = authenticated_client.get(
         url_for(
-            "deliver_grant_funding.add_form",
+            "developers.add_form",
             grant_id=section.collection_schema.grant.id,
             schema_id=section.collection_schema.id,
             section_id=section.id,
@@ -212,7 +210,7 @@ def test_create_form_get(authenticated_client, factories, templates_rendered):
 
     result = authenticated_client.get(
         url_for(
-            "deliver_grant_funding.add_form",
+            "developers.add_form",
             grant_id=section.collection_schema.grant.id,
             schema_id=section.collection_schema.id,
             section_id=section.id,
@@ -230,7 +228,7 @@ def test_create_form_post(authenticated_client, factories, db_session):
     form_form = FormForm(title="My test form")
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.add_form",
+            "developers.add_form",
             grant_id=section.collection_schema.grant.id,
             schema_id=section.collection_schema.id,
             section_id=section.id,
@@ -240,7 +238,7 @@ def test_create_form_post(authenticated_client, factories, db_session):
     )
     assert result.status_code == 302
     assert result.location == url_for(
-        "deliver_grant_funding.manage_section",
+        "developers.manage_section",
         grant_id=section.collection_schema.grant.id,
         schema_id=section.collection_schema.id,
         section_id=section.id,
@@ -256,7 +254,7 @@ def test_create_form_post_duplicate_name(authenticated_client, factories, db_ses
     form_form = FormForm(title="My test form")
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.add_form",
+            "developers.add_form",
             grant_id=section.collection_schema.grant.id,
             schema_id=section.collection_schema.id,
             section_id=section.id,
@@ -278,7 +276,7 @@ def test_move_section(authenticated_client, factories, db_session):
 
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.move_section",
+            "developers.move_section",
             grant_id=schema.grant.id,
             schema_id=schema.id,
             section_id=section1.id,
@@ -287,7 +285,7 @@ def test_move_section(authenticated_client, factories, db_session):
     )
     assert result.status_code == 302
     assert result.location == url_for(
-        "deliver_grant_funding.list_sections",
+        "developers.list_sections",
         grant_id=schema.grant.id,
         schema_id=schema.id,
     )
@@ -300,7 +298,7 @@ def test_move_section(authenticated_client, factories, db_session):
 
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.move_section",
+            "developers.move_section",
             grant_id=schema.grant.id,
             schema_id=schema.id,
             section_id=section1.id,
@@ -309,7 +307,7 @@ def test_move_section(authenticated_client, factories, db_session):
     )
     assert result.status_code == 302
     assert result.location == url_for(
-        "deliver_grant_funding.list_sections",
+        "developers.list_sections",
         grant_id=schema.grant.id,
         schema_id=schema.id,
     )
@@ -322,7 +320,7 @@ def test_move_section(authenticated_client, factories, db_session):
 
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.move_section",
+            "developers.move_section",
             grant_id=schema.grant.id,
             schema_id=schema.id,
             section_id=section1.id,
@@ -338,7 +336,7 @@ def test_move_form(authenticated_client, factories, db_session):
     form2 = factories.form.create(section=section, order=1)
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.move_form",
+            "developers.move_form",
             grant_id=section.collection_schema.grant.id,
             schema_id=section.collection_schema.id,
             section_id=section.id,
@@ -348,7 +346,7 @@ def test_move_form(authenticated_client, factories, db_session):
     )
     assert result.status_code == 302
     assert result.location == url_for(
-        "deliver_grant_funding.manage_section",
+        "developers.manage_section",
         grant_id=section.collection_schema.grant.id,
         schema_id=section.collection_schema.id,
         section_id=section.id,
@@ -360,7 +358,7 @@ def test_move_form(authenticated_client, factories, db_session):
     assert form2_from_db.order == 0
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.move_form",
+            "developers.move_form",
             grant_id=section.collection_schema.grant.id,
             schema_id=section.collection_schema.id,
             section_id=section.id,
@@ -370,7 +368,7 @@ def test_move_form(authenticated_client, factories, db_session):
     )
     assert result.status_code == 302
     assert result.location == url_for(
-        "deliver_grant_funding.manage_section",
+        "developers.manage_section",
         grant_id=section.collection_schema.grant.id,
         schema_id=section.collection_schema.id,
         section_id=section.id,
@@ -383,7 +381,7 @@ def test_move_form(authenticated_client, factories, db_session):
 
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.move_form",
+            "developers.move_form",
             grant_id=section.collection_schema.grant.id,
             schema_id=section.collection_schema.id,
             section_id=section.id,
@@ -400,7 +398,7 @@ def test_move_question(authenticated_client, factories, db_session):
     question2 = factories.question.create(form=form, order=1)
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.move_question",
+            "developers.move_question",
             grant_id=form.section.collection_schema.grant.id,
             schema_id=form.section.collection_schema.id,
             section_id=form.section.id,
@@ -411,7 +409,7 @@ def test_move_question(authenticated_client, factories, db_session):
     )
     assert result.status_code == 302
     assert result.location == url_for(
-        "deliver_grant_funding.manage_form",
+        "developers.manage_form",
         grant_id=form.section.collection_schema.grant.id,
         schema_id=form.section.collection_schema.id,
         section_id=form.section.id,
@@ -426,7 +424,7 @@ def test_move_question(authenticated_client, factories, db_session):
 
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.move_question",
+            "developers.move_question",
             grant_id=form.section.collection_schema.grant.id,
             schema_id=form.section.collection_schema.id,
             section_id=form.section.id,
@@ -437,7 +435,7 @@ def test_move_question(authenticated_client, factories, db_session):
     )
     assert result.status_code == 302
     assert result.location == url_for(
-        "deliver_grant_funding.manage_form",
+        "developers.manage_form",
         grant_id=form.section.collection_schema.grant.id,
         schema_id=form.section.collection_schema.id,
         section_id=form.section.id,
@@ -452,7 +450,7 @@ def test_move_question(authenticated_client, factories, db_session):
 
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.move_question",
+            "developers.move_question",
             grant_id=form.section.collection_schema.grant.id,
             schema_id=form.section.collection_schema.id,
             section_id=form.section.id,
@@ -468,7 +466,7 @@ def test_create_question_choose_type_get(authenticated_client, factories):
     form = factories.form.create()
     result = authenticated_client.get(
         url_for(
-            "deliver_grant_funding.choose_question_type",
+            "developers.choose_question_type",
             grant_id=form.section.collection_schema.grant.id,
             schema_id=form.section.collection_schema.id,
             section_id=form.section.id,
@@ -486,7 +484,7 @@ def test_create_question_choose_type_post(authenticated_client, factories):
     wt_form = QuestionTypeForm(question_data_type=QuestionDataType.TEXT_SINGLE_LINE.name)
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.choose_question_type",
+            "developers.choose_question_type",
             grant_id=form.section.collection_schema.grant.id,
             schema_id=form.section.collection_schema.id,
             section_id=form.section.id,
@@ -497,7 +495,7 @@ def test_create_question_choose_type_post(authenticated_client, factories):
     )
     assert result.status_code == 302
     assert result.location == url_for(
-        "deliver_grant_funding.add_question",
+        "developers.add_question",
         grant_id=form.section.collection_schema.grant.id,
         schema_id=form.section.collection_schema.id,
         section_id=form.section.id,
@@ -511,7 +509,7 @@ def test_create_question_choose_type_post_error(authenticated_client, factories)
     wt_form = QuestionTypeForm()
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.choose_question_type",
+            "developers.choose_question_type",
             grant_id=form.section.collection_schema.grant.id,
             schema_id=form.section.collection_schema.id,
             section_id=form.section.id,
@@ -530,7 +528,7 @@ def test_add_text_question_get(authenticated_client, factories):
     form = factories.form.create()
     result = authenticated_client.get(
         url_for(
-            "deliver_grant_funding.add_question",
+            "developers.add_question",
             grant_id=form.section.collection_schema.grant.id,
             schema_id=form.section.collection_schema.id,
             section_id=form.section.id,
@@ -554,7 +552,7 @@ def test_add_text_question_post(authenticated_client, factories, db_session):
     wt_form = QuestionForm(text="Text question 1", hint="some hint text", name="question 1")
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.add_question",
+            "developers.add_question",
             grant_id=form.section.collection_schema.grant.id,
             schema_id=form.section.collection_schema.id,
             section_id=form.section.id,
@@ -565,7 +563,7 @@ def test_add_text_question_post(authenticated_client, factories, db_session):
     )
     assert result.status_code == 302
     assert result.location == url_for(
-        "deliver_grant_funding.manage_form",
+        "developers.manage_form",
         grant_id=form.section.collection_schema.grant.id,
         schema_id=form.section.collection_schema.id,
         section_id=form.section.id,
@@ -584,7 +582,7 @@ def test_add_text_question_post_duplicate_text(authenticated_client, factories, 
     wt_form = QuestionForm(text="duplicate text", hint="some hint text", name="question 1")
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.add_question",
+            "developers.add_question",
             grant_id=form.section.collection_schema.grant.id,
             schema_id=form.section.collection_schema.id,
             section_id=form.section.id,
@@ -606,7 +604,7 @@ def test_edit_question_get(authenticated_client, factories):
     question = factories.question.create(form=form, text="Test Question", hint="Test Hint", name="Test Question Name")
     result = authenticated_client.get(
         url_for(
-            "deliver_grant_funding.edit_question",
+            "developers.edit_question",
             grant_id=form.section.collection_schema.grant.id,
             schema_id=form.section.collection_schema.id,
             section_id=form.section.id,
@@ -626,7 +624,7 @@ def test_edit_question_post(authenticated_client, factories, db_session):
     wt_form = QuestionForm(text="Updated Question", hint="Updated Hint", name="Updated Question Name")
     result = authenticated_client.post(
         url_for(
-            "deliver_grant_funding.edit_question",
+            "developers.edit_question",
             grant_id=form.section.collection_schema.grant.id,
             schema_id=form.section.collection_schema.id,
             section_id=form.section.id,
@@ -637,7 +635,7 @@ def test_edit_question_post(authenticated_client, factories, db_session):
     )
     assert result.status_code == 302
     assert result.location == url_for(
-        "deliver_grant_funding.manage_form",
+        "developers.manage_form",
         grant_id=form.section.collection_schema.grant.id,
         schema_id=form.section.collection_schema.id,
         section_id=form.section.id,
